+++ src/main/java/org/mule/extension/ftp/api/proxy/HttpTunnelProxy.java
@@ -7,8 +7,20 @@
 package org.mule.extension.ftp.api.proxy;
 
 import org.mule.runtime.api.tls.TlsContextFactory;
+import org.mule.runtime.extension.api.annotation.Expression;
+import org.mule.runtime.extension.api.annotation.param.Optional;
+import org.mule.runtime.extension.api.annotation.param.Parameter;
+import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
+import org.mule.runtime.extension.api.annotation.param.display.Password;
 import org.mule.runtime.extension.api.annotation.param.display.Summary;
 import org.mule.sdk.api.annotation.semantics.connectivity.ConfiguresProxy;
+import org.mule.sdk.api.annotation.semantics.connectivity.Host;
+import org.mule.sdk.api.annotation.semantics.connectivity.Port;
+import org.mule.sdk.api.annotation.semantics.security.Username;
+
+import java.util.Objects;
+
+import static org.mule.runtime.api.meta.ExpressionSupport.SUPPORTED;
 
 /**
  * Groups FTP Proxy connection parameters
@@ -18,11 +30,103 @@ import org.mule.sdk.api.annotation.semantics.connectivity.ConfiguresProxy;
 
 @Summary("The FTP Proxy Settings")
 @ConfiguresProxy
-public class HttpTunnelProxy extends Proxy implements ProxySettings {
+public class HttpTunnelProxy implements ProxySettings {
+
+  /**
+   * The FTP Proxy server host, such as www.mulesoft.com, localhost, or 192.168.0.1, etc
+   */
+  @Parameter
+  @DisplayName("Proxy host")
+  @Expression(SUPPORTED)
+  @Optional(defaultValue = "")
+  @Host
+  protected String host;
+
+  /**
+   * The port number of the FTP Proxy server to connect
+   */
+  @Parameter
+  @DisplayName("Proxy port")
+  @Expression(SUPPORTED)
+  @Optional(defaultValue = "")
+  @Port
+  protected int port = 3128;
+
+  /**
+   * Username for the FTP Proxy Server. Required if the Proxy server is authenticated.
+   */
+  @Parameter
+  @DisplayName("Proxy username")
+  @Expression(SUPPORTED)
+  @Optional(defaultValue = "")
+  @Username
+  protected String username;
+
+  /**
+   * Password for the FTP Proxy Server. Required if the Proxy server is authenticated.
+   */
+  @Parameter
+  @DisplayName("Proxy password")
+  @Expression(SUPPORTED)
+  @Optional(defaultValue = "")
+  @Password
+  protected String password;
+
+  @Override
+  public String getHost() {
+    return host;
+  }
+
+  public void setHost(String host) {
+    this.host = host;
+  }
+
+  @Override
+  public int getPort() {
+    return port;
+  }
+
+  public void setPort(int port) {
+    this.port = port;
+  }
+
+  @Override
+  public String getUsername() {
+    return username;
+  }
+
+  public void setUsername(String username) {
+    this.username = username;
+  }
+
+  @Override
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
 
   @Override
   public TlsContextFactory getTlsContextFactory() {
     return null;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o)
+      return true;
+    if (o == null || getClass() != o.getClass())
+      return false;
+    HttpTunnelProxy that = (HttpTunnelProxy) o;
+    return port == that.port && Objects.equals(host, that.host) && Objects.equals(username, that.username)
+        && Objects.equals(password, that.password);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(host, port, username, password);
   }
 
 }
