+++ src/test/resources/dockercompose/docker-compose.yml
@@ -10,7 +10,7 @@ services:
       - ./squid.conf:/etc/squid/squid.conf
       - ./myCA.pem:/etc/squid/certs/key.pem
   ftp:
-    image: "stilliard/pure-ftpd"
+    image: "fauria/vsftpd"
     container_name: "ftp"
     ports:
       - "21:21"
@@ -18,7 +18,5 @@ services:
     environment:
       PUBLICHOST: localhost
       FTP_USER_HOME: /home/guest
-      FTP_USER_NAME: guest
-      FTP_USER_PASS: guest
-      FTP_MAX_CLIENTS: 50
-      FTP_MAX_CONNECTIONS: 50
+      FTP_USER: guest
+      FTP_PASS: guest
